// TestUI/src/app/models/weather.ts
  export enum WeatherClass {
    CYAN = 'cyan',
    GREEN = 'green',
    ORANGE = 'orange',
    RED = 'red'
  }
  
  export enum WeatherDescription {
    FREEZING = 'Freezing',
    BRACING = 'Bracing',
    CHILLY = 'Chilly',
    MILD = 'Mild',
    BALMY = 'Balmy',
    COOL = 'Cool',
    WARM = 'Warm',
    HOT = 'Hot',
    SWELTERING = 'Sweltering',
    SCORCHING = 'Scorching',
  }